# openapi/okoscope-v1.yaml
openapi: 3.1.0
info:
  title: Okoscope Web API
  version: 1.0.0
  description: Stable API contract for the separately deployed Okoscope Web UI. Timestamps are RFC 3339 UTC values.
servers:
  - url: /
security:
  - sessionAuth: []
paths:
...
  /api/v1/projects/{project_id}/webhook-destinations:
    parameters: [{ $ref: '#/components/parameters/ProjectId' }]
    get:
      operationId: listWebhookDestinations
      responses: { '200': { $ref: '#/components/responses/WebhookDestinationList' } }
    post:
      operationId: createWebhookDestination
      requestBody: { $ref: '#/components/requestBodies/CreateWebhookDestination' }
      responses: { '201': { $ref: '#/components/responses/DestinationWithSecret' } }
  /api/v1/projects/{project_id}/webhook-destinations/{destination_id}:
    parameters:
      [
        { $ref: '#/components/parameters/ProjectId' },
        { $ref: '#/components/parameters/DestinationId' },
      ]
    get:
      operationId: getWebhookDestination
      responses: { '200': { $ref: '#/components/responses/WebhookDestination' } }
    patch:
      operationId: updateWebhookDestination
      requestBody: { $ref: '#/components/requestBodies/UpdateWebhookDestination' }
      responses: { '200': { $ref: '#/components/responses/WebhookDestination' } }
  /api/v1/projects/{project_id}/webhook-destinations/{destination_id}/disable:
    parameters:
      [
        { $ref: '#/components/parameters/ProjectId' },
        { $ref: '#/components/parameters/DestinationId' },
      ]
    post:
      operationId: disableWebhookDestination
      responses: { '200': { $ref: '#/components/responses/WebhookDestination' } }
  /api/v1/projects/{project_id}/webhook-destinations/{destination_id}/rotate-secret:
    parameters:
      [
        { $ref: '#/components/parameters/ProjectId' },
        { $ref: '#/components/parameters/DestinationId' },
      ]
    post:
      operationId: rotateWebhookSecret
      responses: { '200': { $ref: '#/components/responses/DestinationWithSecret' } }
  /api/v1/projects/{project_id}/webhook-destinations/{destination_id}/test:
    parameters:
      [
        { $ref: '#/components/parameters/ProjectId' },
        { $ref: '#/components/parameters/DestinationId' },
      ]
    post:
      operationId: testWebhookDestination
      responses: { '200': { $ref: '#/components/responses/DeliverySummary' } }
...
  /api/v1/projects/{project_id}/notification-recovery-operations:
    parameters: [{ $ref: '#/components/parameters/ProjectId' }]
    get:
      operationId: listNotificationRecoveryOperations
      parameters:
        - {
            name: command_type,
            in: query,
            schema: { $ref: '#/components/schemas/RecoveryCommandType' },
          }
        - { $ref: '#/components/parameters/Cursor' }
        - { $ref: '#/components/parameters/Limit' }
      responses: { '200': { $ref: '#/components/responses/RecoveryOperationPage' } }
  /api/v1/projects/{project_id}/notification-recovery-operations/{operation_id}:
    parameters:
      [
        { $ref: '#/components/parameters/ProjectId' },
        { $ref: '#/components/parameters/OperationId' },
      ]
    get:
      operationId: getNotificationRecoveryOperation
      responses: { '200': { $ref: '#/components/responses/RecoveryOperationDetail' } }
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: okoscope_session
      description: Opaque HttpOnly browser session; clients send requests with credentials enabled.
  parameters:
...
    Cursor:
      {
        name: cursor,
        in: query,
        schema: { type: string, format: uuid },
        description: Opaque cursor scoped to the authenticated collection.,
      }
...
    Limit:
      { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 50 } }
    ProjectId:
      { name: project_id, in: path, required: true, schema: { type: string, format: uuid } }
...
    DestinationId:
      { name: destination_id, in: path, required: true, schema: { type: string, format: uuid } }
...
    OperationId:
      { name: operation_id, in: path, required: true, schema: { type: string, format: uuid } }
...
  headers:
    RequestId: { schema: { type: string }, description: Correlation identifier for the request. }
...
  schemas:
...
    WebhookDestination:
      type: object
      additionalProperties: false
      required:
        [
          id,
          project_id,
          name,
          url,
          enabled,
          deliver_backfill,
          revision,
          created_at,
          updated_at,
          disabled_at,
        ]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        project_id: { $ref: '#/components/schemas/Uuid' }
        name: { type: string }
        url: { type: string, format: uri }
        enabled: { type: boolean }
        deliver_backfill: { type: boolean }
        revision: { type: integer, format: int64 }
        created_at: { $ref: '#/components/schemas/Timestamp' }
        updated_at: { $ref: '#/components/schemas/Timestamp' }
        disabled_at: { $ref: '#/components/schemas/NullableTimestamp' }
    DestinationWithSecret:
      unevaluatedProperties: false
      allOf:
        - { $ref: '#/components/schemas/WebhookDestination' }
        - type: object
          required: [secret]
          properties: { secret: { type: string } }
    CreateWebhookDestination:
      type: object
      additionalProperties: false
      required: [name, url]
      properties:
        name: { type: string }
        url: { type: string, format: uri }
        deliver_backfill: { type: boolean, description: Defaults to false when omitted. }
    UpdateWebhookDestination:
      type: object
      additionalProperties: false
      required: [revision]
      properties:
        name: { type: string }
        url: { type: string, format: uri }
        deliver_backfill: { type: boolean }
        enabled: { type: boolean }
        revision: { type: integer, format: int64 }
    DeliverySummary:
      type: object
      additionalProperties: false
      required:
        [
          id,
          project_id,
          destination_id,
          outbox_message_id,
          origin,
          source,
          event_name,
          semantic_metadata,
          destination,
          status,
          available_at,
          next_attempt_at,
          recovery_generation,
          attempt_count,
          total_attempt_count,
          max_attempts,
          last_error_class,
          terminal_reason,
          created_at,
          updated_at,
          terminal_at,
          retry_allowed,
          cancel_allowed,
          last_recovery_operation_id,
        ]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        project_id: { $ref: '#/components/schemas/Uuid' }
        destination_id: { $ref: '#/components/schemas/Uuid' }
        outbox_message_id: { $ref: '#/components/schemas/NullableUuid' }
        origin: { type: string }
        source: { type: string }
        event_name: { type: string }
        semantic_metadata: { $ref: '#/components/schemas/NullableDeliverySemanticMetadata' }
        destination: { $ref: '#/components/schemas/SafeDestination' }
        status:
          { type: string, enum: [pending, in_flight, succeeded, failed, suppressed, cancelled] }
        available_at: { $ref: '#/components/schemas/Timestamp' }
        next_attempt_at: { $ref: '#/components/schemas/NullableTimestamp' }
        recovery_generation: { type: integer, format: int32, minimum: 0 }
        attempt_count: { type: integer, format: int32 }
        total_attempt_count: { type: integer, format: int64, minimum: 0 }
        max_attempts: { type: integer, format: int32 }
        last_error_class: { type: [string, 'null'], maxLength: 100 }
        terminal_reason: { type: [string, 'null'], maxLength: 100 }
        created_at: { $ref: '#/components/schemas/Timestamp' }
        updated_at: { $ref: '#/components/schemas/Timestamp' }
        terminal_at: { $ref: '#/components/schemas/NullableTimestamp' }
        retry_allowed: { type: boolean }
        cancel_allowed: { type: boolean }
        last_recovery_operation_id: { $ref: '#/components/schemas/NullableUuid' }
    SafeDestination:
      type: object
      additionalProperties: false
      description: Destination identity and operational state without URL, credentials, or signing material.
      required: [id, name, enabled]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        name: { type: string, maxLength: 200 }
        enabled: { type: boolean }
    DeliverySemanticMetadata:
      type: object
      additionalProperties: false
      required: [application_id, group_id, event_kind]
      properties:
        application_id: { $ref: '#/components/schemas/NullableUuid' }
        group_id: { $ref: '#/components/schemas/NullableUuid' }
        event_kind: { type: [string, 'null'], maxLength: 100 }
    NullableDeliverySemanticMetadata:
      oneOf:
        - { $ref: '#/components/schemas/DeliverySemanticMetadata' }
        - { type: 'null' }
...
    RecoveryCommandType: { type: string, enum: [retry, cancel, bulk_retry] }
...
    RecoveryOperationSummary:
      type: object
      additionalProperties: false
      required:
        [
          id,
          project_id,
          command_type,
          target_delivery_id,
          actor_kind,
          actor_id,
          request_id,
          outcome,
          selected_count,
          retried_count,
          cancelled_count,
          skipped_count,
          remaining_count,
          created_at,
          completed_at,
        ]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        project_id: { $ref: '#/components/schemas/Uuid' }
        command_type: { $ref: '#/components/schemas/RecoveryCommandType' }
        target_delivery_id: { $ref: '#/components/schemas/NullableUuid' }
        actor_kind: { type: string, enum: [api_credential] }
        actor_id: { $ref: '#/components/schemas/Uuid' }
        request_id: { type: string, maxLength: 200 }
        outcome: { type: string, enum: [completed, conflict] }
        selected_count: { type: integer, format: int32, minimum: 0 }
        retried_count: { type: integer, format: int32, minimum: 0 }
        cancelled_count: { type: integer, format: int32, minimum: 0 }
        skipped_count: { type: integer, format: int32, minimum: 0 }
        remaining_count: { type: integer, format: int32, minimum: 0 }
        created_at: { $ref: '#/components/schemas/Timestamp' }
        completed_at: { $ref: '#/components/schemas/Timestamp' }
    RecoveryOperationDelivery:
      type: object
      additionalProperties: false
      required: [delivery_id, recovery_generation, action, created_at]
      properties:
        delivery_id: { $ref: '#/components/schemas/Uuid' }
        recovery_generation: { type: integer, format: int32, minimum: 0 }
        action: { type: string, enum: [retried, cancelled, skipped] }
        created_at: { $ref: '#/components/schemas/Timestamp' }
    RecoveryOperationPage:
      type: object
      additionalProperties: false
      required: [items, next_cursor]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/RecoveryOperationSummary' } }
        next_cursor: { $ref: '#/components/schemas/NullableUuid' }
    RecoveryOperationDetail:
      unevaluatedProperties: false
      allOf:
        - { $ref: '#/components/schemas/RecoveryOperationSummary' }
        - type: object
          required: [affected_deliveries]
          properties:
            affected_deliveries:
              {
                type: array,
                maxItems: 200,
                items: { $ref: '#/components/schemas/RecoveryOperationDelivery' },
              }
...
    Uuid: { type: string, format: uuid }
    NullableUuid: { type: [string, 'null'], format: uuid }
...
    Timestamp: { type: string, format: date-time, example: '2026-08-17T12:00:00Z' }
    NullableTimestamp: { type: [string, 'null'], format: date-time }
  requestBodies:
...
    CreateWebhookDestination:
      required: true
      content:
        { application/json: { schema: { $ref: '#/components/schemas/CreateWebhookDestination' } } }
    UpdateWebhookDestination:
      required: true
      content:
        { application/json: { schema: { $ref: '#/components/schemas/UpdateWebhookDestination' } } }
...
  responses:
...
    WebhookDestination:
      {
        description: Webhook destination,
        headers: { X-Request-Id: { $ref: '#/components/headers/RequestId' } },
        content:
          { application/json: { schema: { $ref: '#/components/schemas/WebhookDestination' } } },
      }
    WebhookDestinationList:
      {
        description: Webhook destinations,
        headers: { X-Request-Id: { $ref: '#/components/headers/RequestId' } },
        content:
          {
            application/json:
              {
                schema: { type: array, items: { $ref: '#/components/schemas/WebhookDestination' } },
              },
          },
      }
    DestinationWithSecret:
      {
        description: Webhook destination with one-time secret,
        headers: { X-Request-Id: { $ref: '#/components/headers/RequestId' } },
        content:
          { application/json: { schema: { $ref: '#/components/schemas/DestinationWithSecret' } } },
      }
    DeliverySummary:
      {
        description: Notification delivery summary,
        headers: { X-Request-Id: { $ref: '#/components/headers/RequestId' } },
        content: { application/json: { schema: { $ref: '#/components/schemas/DeliverySummary' } } },
      }
...
    RecoveryOperationPage:
      {
        description: Recovery operation page,
        headers: { X-Request-Id: { $ref: '#/components/headers/RequestId' } },
        content:
          { application/json: { schema: { $ref: '#/components/schemas/RecoveryOperationPage' } } },
      }
    RecoveryOperationDetail:
      {
        description: Recovery operation detail,
        headers: { X-Request-Id: { $ref: '#/components/headers/RequestId' } },
        content:
          {
            application/json: { schema: { $ref: '#/components/schemas/RecoveryOperationDetail' } },
          },
      }
...